Change interface's properties to save space 8

+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/Triggers/Interfaces/IHasTargetItemID.cs
@@ -13,6 +13,6 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.Triggers.Interfaces
     {
         /// <summary>The target Item ID of the trigger.</summary>
         [ObjectStringMappable(ObjectParameter.ItemID)]
-        int TargetItemID { get; set; }
+        short TargetItemID { get; set; }
     }
 }
